Revert "LPS-169605 Initialize CheckNotificationQueueEntryMessageListener dynamic in LifecycleListener"
