+++ modules/apps/notification/notification-service/src/main/java/com/liferay/notification/internal/messaging/CheckNotificationQueueEntryMessageListener.java
@@ -15,66 +15,41 @@
 package com.liferay.notification.internal.messaging;
 
 import com.liferay.notification.constants.NotificationConstants;
-import com.liferay.notification.internal.configuration.NotificationQueueConfiguration;
 import com.liferay.notification.service.NotificationQueueEntryLocalService;
 import com.liferay.notification.type.NotificationType;
 import com.liferay.notification.type.NotificationTypeServiceTracker;
-import com.liferay.petra.string.StringBundler;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.exception.SystemException;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.messaging.BaseMessageListener;
 import com.liferay.portal.kernel.messaging.DestinationNames;
 import com.liferay.portal.kernel.messaging.Message;
-import com.liferay.portal.kernel.module.configuration.ConfigurationException;
-import com.liferay.portal.kernel.module.configuration.ConfigurationProvider;
+import com.liferay.portal.kernel.module.framework.ModuleServiceLifecycle;
 import com.liferay.portal.kernel.scheduler.SchedulerEngineHelper;
 import com.liferay.portal.kernel.scheduler.SchedulerEntryImpl;
-import com.liferay.portal.kernel.scheduler.SchedulerException;
-import com.liferay.portal.kernel.scheduler.StorageType;
 import com.liferay.portal.kernel.scheduler.TimeUnit;
 import com.liferay.portal.kernel.scheduler.TriggerFactory;
-import com.liferay.portal.kernel.settings.CompanyServiceSettingsLocator;
-import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
+import com.liferay.portal.kernel.scheduler.messaging.SchedulerResponse;
+import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
+import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.Time;
 
 import java.util.Date;
 
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
+import org.osgi.service.component.annotations.Reference;
+
 /**
  * @author Gustavo Lima
  */
+@Component(service = {})
 public class CheckNotificationQueueEntryMessageListener
 	extends BaseMessageListener {
 
-	public CheckNotificationQueueEntryMessageListener(
-		long companyId, ConfigurationProvider configurationProvider,
-		NotificationQueueEntryLocalService notificationQueueEntryLocalService,
-		NotificationTypeServiceTracker notificationTypeServiceTracker,
-		SchedulerEngineHelper schedulerEngineHelper,
-		TriggerFactory triggerFactory) {
+	@Activate
+	protected void activate() {
+		Class<?> clazz = getClass();
 
-		_configurationProvider = configurationProvider;
-		_notificationQueueEntryLocalService =
-			notificationQueueEntryLocalService;
-		_notificationTypeServiceTracker = notificationTypeServiceTracker;
-		_schedulerEngineHelper = schedulerEngineHelper;
-		_triggerFactory = triggerFactory;
-
-		String className = _getClassName(companyId);
-
-		try {
-			_configurationProvider.saveCompanyConfiguration(
-				NotificationQueueConfiguration.class, companyId,
-				HashMapDictionaryBuilder.<String, Object>put(
-					"checkInterval", NotificationConstants.CHECK_INTERVAL
-				).put(
-					"deleteInterval", NotificationConstants.DELETE_INTERVAL
-				).build());
-		}
-		catch (ConfigurationException configurationException) {
-			throw new RuntimeException(configurationException);
-		}
+		String className = clazz.getName();
 
 		_schedulerEngineHelper.register(
 			this,
@@ -86,71 +61,23 @@ public class CheckNotificationQueueEntryMessageListener
 			DestinationNames.SCHEDULER_DISPATCH);
 	}
 
+	@Deactivate
+	protected void deactivate() {
+		_schedulerEngineHelper.unregister(this);
+	}
+
 	@Override
 	protected void doReceive(Message message) throws Exception {
-		synchronized (this) {
-			NotificationType notificationType =
-				_notificationTypeServiceTracker.getNotificationType(
-					NotificationConstants.TYPE_EMAIL);
+		NotificationType notificationType =
+			_notificationTypeServiceTracker.getNotificationType(
+				NotificationConstants.TYPE_EMAIL);
 
-			long companyId = message.getLong("companyId");
+		long companyId = CompanyThreadLocal.getCompanyId();
 
-			notificationType.sendUnsentNotifications(companyId);
+		notificationType.sendUnsentNotifications();
 
-			NotificationQueueConfiguration notificationQueueConfiguration =
-				_getNotificationQueueConfiguration(companyId);
-
-			long deleteInterval =
-				notificationQueueConfiguration.deleteInterval() * Time.MINUTE;
-
-			_notificationQueueEntryLocalService.deleteNotificationQueueEntries(
-				companyId,
-				new Date(System.currentTimeMillis() - deleteInterval));
-
-			String className = _getClassName(companyId);
-
-			try {
-				_schedulerEngineHelper.update(
-					_triggerFactory.createTrigger(
-						className, className, null, null,
-						notificationQueueConfiguration.checkInterval(),
-						TimeUnit.MINUTE),
-					StorageType.MEMORY_CLUSTERED);
-			}
-			catch (SchedulerException schedulerException) {
-				if (_log.isWarnEnabled()) {
-					_log.error(
-						"Unable to update trigger for memory clustered job",
-						schedulerException);
-				}
-			}
-		}
-	}
-
-	private String _getClassName(long companyId) {
-		Class<?> clazz = getClass();
-
-		return StringBundler.concat(
-			clazz.getName(), StringPool.POUND, companyId);
-	}
-
-	private NotificationQueueConfiguration _getNotificationQueueConfiguration(
-		long companyId) {
-
-		try {
-			return _configurationProvider.getConfiguration(
-				NotificationQueueConfiguration.class,
-				new CompanyServiceSettingsLocator(
-					companyId, NotificationConstants.RESOURCE_NAME,
-					NotificationQueueConfiguration.class.getName()));
-		}
-		catch (ConfigurationException configurationException) {
-			if (_log.isDebugEnabled()) {
-				_log.debug(configurationException);
-			}
-
-			throw new SystemException(configurationException);
-		}
+		_notificationQueueEntryLocalService.deleteNotificationQueueEntries(
+			new Date(System.currentTimeMillis() - Time.MONTH));
 	}
 
 	private static final int _CHECK_INTERVAL = 15;
@@ -158,12 +85,17 @@ public class CheckNotificationQueueEntryMessageListener
 	private static final Log _log = LogFactoryUtil.getLog(
 		CheckNotificationQueueEntryMessageListener.class);
 
-	private final ConfigurationProvider _configurationProvider;
-	private volatile NotificationQueueEntryLocalService
+	@Reference
+	private NotificationQueueEntryLocalService
 		_notificationQueueEntryLocalService;
-	private volatile NotificationTypeServiceTracker
-		_notificationTypeServiceTracker;
-	private volatile SchedulerEngineHelper _schedulerEngineHelper;
-	private volatile TriggerFactory _triggerFactory;
+
+	@Reference
+	private NotificationTypeServiceTracker _notificationTypeServiceTracker;
+
+	@Reference
+	private SchedulerEngineHelper _schedulerEngineHelper;
+
+	@Reference
+	private TriggerFactory _triggerFactory;
 
 }
